backend/requirements.txt
beautifulsoup4>=4.12.0
chromadb>=0.4.0
fastapi>=0.115.0
httpx>=0.27.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
lxml>=5.0.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
pypdf>=4.0.0
pytest>=8.0.0
pytest-asyncio
python-docx>=1.1.0
python-dotenv
python-multipart>=0.0.9
PyJWT[crypto]>=2.8.0
redis>=5.0.0
respx>=0.21.0
sentence-transformers>=2.0.0
uvicorn[standard]>=0.32.0
